backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
django-cors-headers==4.3.0
python-dateutil==2.8.2
drf-yasg==1.21.7
